feat: 🎸 update variant changes for test bug fix
