+++ server/src/internal/products/actions/index.ts
@@ -1,5 +1,7 @@
 import { getFreeDefaultProductByGroup } from "@/internal/products/actions/getFreeDefaultProductByGroup";
+import { updateVariant } from "@/internal/products/actions/updateVariant";
 
 export const productActions = {
 	getFreeDefaultByGroup: getFreeDefaultProductByGroup,
+	updateVariant,
 };
